third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/css-flexbox/abspos/flex-abspos-staticpos-justify-content-007-expected.txt

This is a testharness.js-based test.
PASS .container > div 1
FAIL .container > div 2 assert_equals: 
<div class="container" style="justify-content: space-between"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetTop expected 1 but got 5
PASS .container > div 3
PASS .container > div 4
PASS .container > div 5
PASS .container > div 6
FAIL .container > div 7 assert_equals: 
<div class="container" style="justify-content: start"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetTop expected 1 but got 5
FAIL .container > div 8 assert_equals: 
<div class="container" style="justify-content: end"><div data-offset-x="2" data-offset-y="5"></div></div>
offsetTop expected 5 but got 1
PASS .container > div 9
PASS .container > div 10
FAIL .container > div 11 assert_equals: 
<div class="container" style="justify-content: left"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetTop expected 1 but got 5
FAIL .container > div 12 assert_equals: 
<div class="container" style="justify-content: right"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetTop expected 1 but got 5
PASS .container > div 13
FAIL .container > div 14 assert_equals: 
<div class="container" style="justify-content: space-between"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetTop expected 1 but got -3
PASS .container > div 15
PASS .container > div 16
PASS .container > div 17
PASS .container > div 18
FAIL .container > div 19 assert_equals: 
<div class="container" style="justify-content: start"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetTop expected 1 but got -3
FAIL .container > div 20 assert_equals: 
<div class="container" style="justify-content: end"><div data-offset-x="2" data-offset-y="-3"></div></div>
offsetTop expected -3 but got 1
PASS .container > div 21
PASS .container > div 22
FAIL .container > div 23 assert_equals: 
<div class="container" style="justify-content: left"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetTop expected 1 but got -3
FAIL .container > div 24 assert_equals: 
<div class="container" style="justify-content: right"><div data-offset-x="2" data-offset-y="1"></div></div>
offsetTop expected 1 but got -3
Harness: the test ran to completion.
